Moved cloud_merge headers into include/cloud_merge
