+++ cloud_merge/include/cloud_merge/cloud_merge_node.h
