+++ cloud_merge/src/cloud_merge.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include "cloud_merge.h"
+#include "cloud_merge/cloud_merge.h"
 
 
 template class CloudMerge<pcl::PointXYZRGB>;
