+++ cloud_merge/src/cloud_merge_node.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
 
-#include "cloud_merge_node.h"
+#include "cloud_merge/cloud_merge_node.h"
 
 template class CloudMergeNode<pcl::PointXYZRGB>;
